udemy_django_api/requirements.txt
Django==2.2.3
djangorestframework==3.6.2
pytz==2019.1
sqlparse==0.3.0
